% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deforecast.R
\name{deforecast}
\alias{deforecast}
\title{Run deforestation simulation}
\usage{
deforecast(
  xdim = 1000,
  ydim = 1000,
  explicit = FALSE,
  ppl_loc = NULL,
  ppl_scaling = 1,
  ppl_n = 1000,
  budget = 1000,
  p_loc = NULL,
  np_loc = NULL,
  mean_age = 50,
  sd_age = 25,
  p_n = 5000,
  np_n = 5000,
  travel_cost = 100,
  management_cost = 100,
  value = 500,
  mobility = 10,
  k = 1,
  r = 0.01,
  maturity = 10,
  max_age = 100,
  dispersion = 5,
  years = 10
)
}
\arguments{
\item{xdim}{Size of landscape on x axis.}

\item{ydim}{Size of landscape on y axis}

\item{explicit}{If FALSE, places people and trees randomly on the landscape. If TRUE takes input data.}

\item{ppl_loc}{Only if explicit_people is TRUE. A 4 column matrix with id, x/y locations and budget.}

\item{ppl_scaling}{Only if explicit_people is TRUE. A number between 0 and 1 to scale down the # of people on the map.}

\item{ppl_n}{Only if explicit_people is FALSE. The number of people on the landscape.}

\item{budget}{The starting resources of people on the landscape.}

\item{p_loc}{Only if explicit_trees is TRUE. A 4 column matrix with ID, x/y location and age of protected trees.}

\item{np_loc}{As p_loc, but for unprotected trees.}

\item{mean_age}{Mean of tree age.}

\item{sd_age}{Standard deviation of tree age.}

\item{p_n}{The number of protected trees on the landscape.}

\item{np_n}{The number of unprotected trees on the landscape.}

\item{travel_cost}{The cost for a person to move one grid cell to fell a tree.}

\item{management_cost}{The cost for a person to fell a tree in a protected area.}

\item{value}{The value of a tree to a person.}

\item{mobility}{The distance that a person moves towards the trees in each year.}

\item{k}{Carrying capacity: forests can't grow beyond k times the size of the landscape.}

\item{r}{The annual rate of growth of the forest.}

\item{maturity}{The age that trees start to produce seeds.}

\item{max_age}{The age that trees stop producing seeds.}

\item{dispersion}{The distance that seeds can travel from the seeding tree.}

\item{years}{The number of years for the simulation}
}
\description{
Run deforestation simulation
}
