% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felling.r
\name{felling}
\alias{felling}
\title{Does the felling (peoples' budgets change and trees cease to exist)}
\usage{
felling(people, trees, value, management_cost, travel_cost, mobility)
}
\arguments{
\item{people}{Location of agents on the landscape}

\item{trees}{Location, age and protected status of trees on the landscape}

\item{value}{The value of a tree to a person.}

\item{management_cost}{The cost for a person to fell a tree in a protected area.}

\item{travel_cost}{The cost for a person to move one grid cell to fell a tree.}

\item{mobility}{The distance that a person moves towards the trees in each year.}
}
\description{
Does the felling (peoples' budgets change and trees cease to exist)
}
