% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpeople.R
\name{get_people}
\alias{get_people}
\title{Gets people from input data}
\usage{
get_people(ppl_loc, ppl_scaling)
}
\arguments{
\item{ppl_loc}{Only if explicit_people is TRUE. A 4 column matrix with id, x/y locations and budget.}

\item{ppl_scaling}{Only if explicit_people is TRUE. A number between 0 and 1 to scale down the # of people on the map.}
}
\description{
Gets people from input data
}
