% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettree.R
\name{get_trees}
\alias{get_trees}
\title{Makes trees}
\usage{
get_trees(p_loc, np_loc, mean_age, sd_age)
}
\arguments{
\item{p_loc}{Only if explicit_trees is TRUE. A 4 column matrix with ID, x/y location and age of protected trees.}

\item{np_loc}{As p_loc, but for unprotected trees.}

\item{mean_age}{Mean of tree age.}

\item{sd_age}{Standard deviation of tree age.}
}
\value{
A 2-member list with protected and unprotected trees.
}
\description{
Makes trees
}
