% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makepeople.r
\name{make_people}
\alias{make_people}
\title{Makes people}
\usage{
make_people(xdim, ydim, ppl_n)
}
\arguments{
\item{xdim}{Size of landscape on x axis.}

\item{ydim}{Size of landscape on y axis}

\item{ppl_n}{Only if explicit_people is FALSE. The number of people on the landscape.}
}
\description{
Makes people
}
