% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketree.r
\name{make_trees}
\alias{make_trees}
\title{Makes trees}
\usage{
make_trees(xdim, ydim, p_n, np_n, mean_age, sd_age)
}
\arguments{
\item{xdim}{Size of landscape on x axis.}

\item{ydim}{Size of landscape on y axis}

\item{p_n}{The number of protected trees on the landscape.}

\item{np_n}{The number of unprotected trees on the landscape.}

\item{mean_age}{Mean of tree age.}

\item{sd_age}{Standard deviation of tree age.}
}
\value{
A 2-member list with protected and unprotected trees.
}
\description{
Makes trees
}
