% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate.R
\name{rate}
\alias{rate}
\title{Does the felling (peoples' budgets change and trees cease to exist)}
\usage{
rate(trees, k, r)
}
\arguments{
\item{trees}{location, age and protected status of trees}

\item{k}{Carrying capacity: forests can't grow beyond k times the size of the landscape.}

\item{r}{The annual rate of growth of the forest.}
}
\description{
Does the felling (peoples' budgets change and trees cease to exist)
}
